Updated TH2 converter, added comments.
Now using matrix plot as it more closely resembles a TH2. This has
both advantages and drawbacks. The last row and column are now properly
rendered. Currently using `matrix plot*` as the no-starred version
inverts the vertical axis. The plot does not permit holes in the matrix,
so the z domain must be large enough to permit all data points. Using
`restrict z to domain*` sets points outside the domain to the limit
encountered. Changes to the view angle no longer make sense. Switching
between `surf` and `matrix plot` requires that the bin coordinates
change from the corner to the center.

Added option string. If the string "surf" is passed the default `matrix
plot*` is swapped with `surf` and the coordinates are adjusted.