Split the TH1 plotting from Write to a static method.
This permits a user to creat the addplot line without the plot header
and footer. In addition it will permit the Write command to be used for
any type of stored plot.

Also, added the Add method for TH2. This downcasts them to TH1 objects
to store them in the hist map. Currently the 2D hist is only stored if
the map is empty as we can not deal with multiple 2D hists.